move getters time related from group.js to state.js
