+++ frontend/model/contracts/group.js
@@ -14,10 +14,10 @@ import {
 } from './voting/constants.js'
 import { paymentStatusType, paymentType, PAYMENT_COMPLETED } from './payments/index.js'
 import * as Errors from '../errors.js'
-import { merge, deepEqualJSONType, omit, uniq } from '~/frontend/utils/giLodash.js'
-import { currentMonthstamp, ISOStringToMonthstamp, compareMonthstamps, prevMonthstamp, dateFromMonthstamp, lastDayOfMonth } from '~/frontend/utils/time.js'
+import { merge, deepEqualJSONType, omit } from '~/frontend/utils/giLodash.js'
+import { currentMonthstamp, ISOStringToMonthstamp, compareMonthstamps } from '~/frontend/utils/time.js'
 import { vueFetchInitKV } from '~/frontend/views/utils/misc.js'
-import incomeDistribution from '~/frontend/utils/distribution/mincome-proportional.js'
+import groupIncomeDistribution from '~/frontend/utils/distribution/group-income-distribution.js'
 import currencies, { saferFloat } from '~/frontend/views/utils/currencies.js'
 import L from '~/frontend/views/utils/translations.js'
 import {
@@ -97,44 +97,6 @@ function initFetchMonthlyPayments ({ meta, state, getters }) {
   const monthlyPayments = vueFetchInitKV(state.paymentsByMonth, monthstamp, initPaymentMonth({ getters }))
   clearOldPayments({ state })
   return monthlyPayments
-}
-
-function groupIncomeDistribution ({ state, getters, monthstamp, adjusted }) {
-  // the monthstamp will always be for the current month. the alternative
-  // is to allow the re-generation of the distribution for previous months,
-  // but that approach requires also storing the historical mincomeAmount
-  // and historical groupProfiles. Since together these change across multiple
-  // locations in the code, it involves less 'code smell' to do it this way.
-  // see historical/group.js for the ugly way of doing it.
-  const mincomeAmount = getters.groupMincomeAmount
-  const groupProfiles = getters.groupProfiles
-  const currentIncomeDistribution = []
-  for (const username in groupProfiles) {
-    const profile = groupProfiles[username]
-    const incomeDetailsType = profile && profile.incomeDetailsType
-    if (incomeDetailsType) {
-      const adjustment = incomeDetailsType === 'incomeAmount' ? 0 : mincomeAmount
-      const amount = adjustment + profile[incomeDetailsType]
-      currentIncomeDistribution.push({
-        name: username,
-        amount: saferFloat(amount)
-      })
-    }
-  }
-  var dist = incomeDistribution(currentIncomeDistribution, mincomeAmount)
-  if (adjusted) {
-    // if this user has already made some payments to other users this
-    // month, we need to take that into account and adjust the distribution.
-    // this will be used by the Payments page to tell how much still
-    // needs to be paid (if it was a partial payment).
-    for (const p of dist) {
-      const alreadyPaid = getters.paymentTotalFromUserToUser(p.from, p.to, monthstamp)
-      // if we "overpaid" because we sent late payments, remove us from consideration
-      p.amount = saferFloat(Math.max(0, p.amount - alreadyPaid))
-    }
-    dist = dist.filter(p => p.amount > 0)
-  }
-  return dist
 }
 
 DefineContract({
@@ -227,235 +189,7 @@ DefineContract({
         return saferFloat(total)
       }
     },
-    ourPayments (state, getters) {
-      const currency = currencies[getters.groupSettings.mincomeCurrency]
-      const ourUsername = getters.ourUsername
-      const cMonthstamp = currentMonthstamp()
-      const pDate = dateFromMonthstamp(cMonthstamp)
-      const dueIn = lastDayOfMonth(pDate)
-      const monthlyPayments = getters.groupMonthlyPayments
-      const allPayments = getters.currentGroupState.payments
 
-      const sent = (() => {
-        const payments = []
-        // QUESTION: Is this only from this month? What about the previous payments from months ago?
-        //           Same question for received.
-        for (const monthstamp of Object.keys(monthlyPayments).sort()) {
-          const { paymentsFrom } = monthlyPayments[monthstamp]
-
-          if (!paymentsFrom) { continue }
-
-          for (const toUser in paymentsFrom[getters.ourUsername]) {
-            for (const paymentHash of paymentsFrom[getters.ourUsername][toUser]) {
-              const { data, meta } = allPayments[paymentHash]
-
-              payments.push({ hash: paymentHash, data, meta })
-            }
-          }
-        }
-        return payments
-      })()
-      const received = (() => {
-        const thisMonthPayments = monthlyPayments[cMonthstamp]
-        const paymentsFrom = thisMonthPayments && thisMonthPayments.paymentsFrom
-        const payments = []
-
-        if (!paymentsFrom) { return [] }
-
-        for (const fromUser in paymentsFrom) {
-          for (const toUser in paymentsFrom[fromUser]) {
-            if (toUser === getters.ourUsername) {
-              for (const paymentHash of paymentsFrom[fromUser][toUser]) {
-                const { data, meta } = allPayments[paymentHash]
-
-                payments.push({ hash: paymentHash, data, meta })
-              }
-            }
-          }
-        }
-
-        return payments
-      })()
-      const todo = (() => {
-        const payments = []
-        const unadjusted = getters.groupIncomeDistribution.filter(p => p.from === ourUsername)
-
-        for (const p of getters.groupIncomeAdjustedDistribution) {
-          if (p.from !== ourUsername) { continue }
-          const existPayment = unadjusted.find(({ to }) => to === p.to) || { amount: 0 }
-          const amount = +currency.displayWithoutCurrency(p.amount)
-          const existingAmount = +currency.displayWithoutCurrency(existPayment.amount)
-
-          if (amount <= 0) { continue }
-
-          const partialAmount = existingAmount - amount
-          const existingPayment = {}
-          if (partialAmount > 0) {
-            // TODO/BUG this only work if the payment is done in 2 parts. if done in >=3 won't work.
-            const sentPartial = sent.find((s) => s.username === p.to && s.amount === partialAmount)
-            if (sentPartial) {
-              existingPayment.hash = sentPartial.hash // Q: Why do we need this hash?
-            }
-          }
-
-          payments.push({
-            ...existingPayment,
-            ...p,
-            total: existingAmount,
-            partial: partialAmount > 0,
-            dueIn
-          })
-        }
-        return payments
-      })()
-      const late = (() => {
-        // const currentDistribution = this.groupIncomeAdjustedDistribution
-        // const { pledgeAmount } = this.ourGroupProfile
-        const pMonthstamp = prevMonthstamp(cMonthstamp)
-        const latePayments = []
-        const pastMonth = monthlyPayments[pMonthstamp]
-        // QUESTION: Is this only from the previous month? What if there are payments from 2-3 months ago?
-        if (pastMonth) {
-          const pDate = dateFromMonthstamp(pMonthstamp)
-          const dueIn = lastDayOfMonth(pDate)
-          const adjusted = getters.groupIncomeAdjustedDistributionForMonth(pMonthstamp)
-
-          // This logic below is wrong (based on cypress tests). Commented for later analysis.
-          /*
-          for (const payment of pastMonth.lastAdjustedDistribution) {
-            if (payment.from === this.ourUsername && payment.amount > 0) {
-              // Let A = the amount we owe from the previous distribution.
-              // Let B = the total we've sent to payment.to from the current
-              //         month's paymentsFrom.
-              // Let C = the total amount we "owe" to payment.to from the
-              //         current month's distribution.
-              // Let D = the amount we're pledging this month
-              // Let E = the amount still unpaid for the previous month's distribution,
-              //         calculated as: C > 0 ? A : A + D - B
-              //
-              // If E > 0, then display a row for the late payment.
-              const A = payment.amount
-              const B = this.paymentTotalFromUserToUser(this.ourUsername, payment.to, cMonthstamp)
-              var C = currentDistribution
-                .filter(a => a.from === payment.from && a.to === payment.to)
-              C = C.length > 0 ? C[0].amount : 0
-              const D = pledgeAmount
-              const E = C > 0 ? A : A + D - B
-              if (E > 0) {
-                latePayments.push({
-                  username: payment.to,
-                  displayName: this.userDisplayName(payment.to),
-                  amount: payment.amount, // TODO: include currency (what if it was changed?)
-                  // partiaL: TODO add this info as it is in this.paymentsTodo
-                  isLate: true,
-                  date: dueIn
-                })
-              }
-            }
-          }
-          */
-
-          // On the other hand this seems to work as expected. "partial" is the only thing missing
-          for (const payment of adjusted) {
-            if (payment.from !== ourUsername) {
-              continue
-            }
-
-            latePayments.push({
-              ...payment,
-              // partial: TODO add this info, but what's the cleanest way to do it?
-              date: dueIn
-            })
-          }
-        }
-        return latePayments
-      })()
-      const toBeReceived = (() => {
-        const unadjusted = getters.groupIncomeDistribution.filter(p => p.to === ourUsername)
-        const payments = []
-        for (const p of getters.groupIncomeAdjustedDistribution) {
-          if (p.to !== ourUsername) { continue }
-
-          const existPayment = unadjusted.find(({ from }) => from === p.from) || { amount: 0 }
-          const amount = +currency.displayWithoutCurrency(p.amount)
-          const existingAmount = +currency.displayWithoutCurrency(existPayment.amount)
-
-          if (amount <= 0) { continue }
-
-          const partialAmount = existingAmount - amount
-          const existingPayment = {}
-          if (partialAmount > 0) {
-            // TODO/BUG this only work if the payment is done in 2 parts. if done in >=3 won't work.
-            const receivePartial = received.find((r) => r.username === p.to && r.amount === partialAmount)
-            if (receivePartial) {
-              existingPayment.hash = receivePartial.hash
-            }
-          }
-          payments.push({
-            ...existingPayment,
-            ...p,
-            total: existingAmount,
-            partial: partialAmount > 0
-          })
-        }
-        return payments
-      })()
-
-      return { sent, todo, late, received, toBeReceived }
-    },
-    ourPaymentsSummary (state, getters) {
-      const { todo, sent, toBeReceived, received } = getters.ourPayments
-      const isCompleted = (p) => p.data.status === PAYMENT_COMPLETED
-      const isPartialCount = (list) => list.filter(p => p.partial).length
-      const getUniqPCount = (users) => {
-        // We need to filter the partial payments already done (sent/received).
-        // E.G. We need to send 4 payments. We've sent 1 full payment and another
-        // in 2 parts. The total must be 2, instead of 3. A quick way to solve this
-        // is by listing all usernames we sent to and count the uniq ones.
-        return uniq(users).length
-      }
-
-      if (getters.ourGroupProfile.incomeDetailsType === 'incomeAmount') {
-        const receivedCompleted = received.filter(isCompleted)
-        const pPartials = isPartialCount(toBeReceived)
-        const pByUser = {
-          toBeReceived: toBeReceived.map(p => p.from),
-          received: received.map(p => p.meta.username)
-        }
-        return {
-          paymentsDone: getUniqPCount(pByUser.received) - pPartials,
-          hasPartials: pPartials > 0,
-          paymentsTotal: getUniqPCount([...pByUser.toBeReceived, ...pByUser.received]),
-          amountDone: receivedCompleted.reduce((total, p) => total + p.data.amount, 0),
-          amountTotal: toBeReceived.reduce((total, p) => total + p.amount, 0) + received.reduce((total, p) => total + p.data.amount, 0)
-        }
-      } else {
-        const sentCompleted = sent.filter(isCompleted)
-        const pPartials = isPartialCount(todo)
-        const pByUser = {
-          todo: todo.map(p => p.to),
-          sent: sent.map(p => p.data.toUser)
-        }
-        return {
-          paymentsDone: getUniqPCount(pByUser.sent) - pPartials,
-          hasPartials: pPartials > 0,
-          paymentsTotal: getUniqPCount([...pByUser.todo, ...pByUser.sent]),
-          amountDone: sentCompleted.reduce((total, p) => total + p.data.amount, 0),
-          amountTotal: todo.reduce((total, p) => total + p.amount, 0) + sent.reduce((total, p) => total + p.data.amount, 0)
-        }
-      }
-    },
-    // used with graphs like those in the dashboard and in the income details modal
-    groupIncomeDistribution (state, getters) {
-      return groupIncomeDistribution({ state, getters, monthstamp: currentMonthstamp() })
-    },
-    // adjusted version of groupIncomeDistribution, used by the payments system
-    groupIncomeAdjustedDistribution (state, getters) {
-      return getters.groupIncomeAdjustedDistributionForMonth(currentMonthstamp())
-    },
-    groupIncomeAdjustedDistributionForMonth (state, getters) {
-      return monthstamp => groupIncomeDistribution({ state, getters, monthstamp, adjusted: true })
-    },
     groupMembersByUsername (state, getters) {
       return Object.keys(getters.groupProfiles)
     },
